Clarify which (arg)min/max index/value is returned
For performance, it's beneficial to avoid guaranteeing a particular
iteration order. For example, the current implementation of `min` may
return any of the minima because the iteration order of
`ArrayBase.fold()` is unspecified. The current implementation of
`argmin` does always return the first minimum (in logical order) since
`ArrayBase.indexed_iter()` always iterates in logical order, but we
may want to optimize the iteration order in the future.