Add -f option when installing tarpaulin in CI
The `tarpaulin` README says "Because of the use of nightly proc-macro
features you'll need to reinstall tarpaulin each time unless you're
keeping to a specific nightly version. If you are keeping to a
specific nightly you can remove the `-f` flag in the example travis
file."

We haven't pinned a nightly version, so we should use the `-f` option.