Set and enforce an msrv of 1.64.0
This crate had an MSRV of 1.64 that was enforced in the CI; however, this was done by way of an explicit pin. Now that Rust comes with an MSRV-aware resolver, we set the `rust-version` explicitly and add a `Cargo.lock` that represents the dependencies needed to support that.

We also add a weekly CI job to ensure that the MSRV still works.